+++ src/manager/ManagerArticle.php
