+++ src/manager/ManagerValide.php
